+++ changelog.md
@@ -3,6 +3,8 @@
 ## Upcoming
 * Rowing boats now gives exhaustion to the player
 * Sweet berry bushes now grow 2x slower in autumn
+* Added a Seasons feature
+  * Has one config option that changes Serene Season's config to disable seasonal crops as it's handled by Plants Growth
 
 ## 3.14.2
 * Requires InsaneLib 1.14.0